refactor(runtime): colocate the shared spawner input contract
