+++ packages/daemon/src/runtime-spawn-context.ts
@@ -2,7 +2,7 @@ import type { AgentRuntimeEventV1, CanonicalEventStore, SessionIdentity } from "
 import type { readDispatchStream } from "./dispatch-stream.ts";
 import type { JsonObject } from "./protocol/json-rpc-types.ts";
 import type { RuntimeAttemptOutcome } from "./runtime-fallback-contract.ts";
-import type { RuntimeSpawnerInput } from "./runtime-spawner.ts";
+import type { RuntimeSpawnerInput } from "./runtime-spawn-types.ts";
 import type { ActiveRuntime, RuntimeBinding, RuntimeAttemptTerminal } from "./runtime-spawn-types.ts";
 import type { launchExitNotification } from "./runtime-spawn-process.ts";
 import type { requiredRuntimeProjection, requiredRuntimeStore } from "./runtime-spawn-process.ts";
